requirements.txt
requests
minio
python-dotenv
